+++ app/models/scenarios/scenario_522.rb
@@ -57,7 +57,6 @@ module Scenarios
           map_data:,
           allied_units:,
           axis_units:,
-          special_rules: ["winter"],
         }
       end