+++ changelog.md
@@ -1,8 +1,10 @@
 # 0.150
 
+- Add cave movement for 420, etc.
 - Add CC moral-size bonus
+- Bug fix: winter/snow movement over frozen water
 - Bug fix: allow crews to abandon immobilized tanks
-- Updated 010, 429
+- Updated 010, 025, 429, 519, 522
 - Added prototype 423
 
 # 0.149